// src/lib/api/v1.d.ts
export interface components {
    schemas: {
// ...
        ProjectDto: {
            id?: string | null;
            name?: string | null;
            /** Format: date-time */
            createdAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}